pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["warepact"]

[project]
name = "warepact"
version = "0.1.0"
description = "The dbt of data contracts — define, enforce, and evolve data contracts"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
keywords = ["data", "contracts", "dbt", "data-quality", "data-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "fastmcp>=0.1",
    "sqlglot>=23.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "croniter>=1.0",
]

[project.optional-dependencies]
snowflake = ["snowflake-connector-python>=3.0"]
bigquery = ["google-cloud-bigquery>=3.0"]
redshift = ["redshift-connector>=2.0"]
duckdb = ["duckdb>=0.10"]
postgres = ["psycopg2-binary>=2.9"]
s3 = ["boto3>=1.26"]
gcs = ["google-cloud-storage>=2.0"]
databricks = ["databricks-sql-connector>=3.0"]
ai = ["anthropic>=0.25"]
all = [
    "snowflake-connector-python>=3.0",
    "google-cloud-bigquery>=3.0",
    "redshift-connector>=2.0",
    "duckdb>=0.10",
    "psycopg2-binary>=2.9",
    "boto3>=1.26",
    "google-cloud-storage>=2.0",
    "databricks-sql-connector>=3.0",
    "anthropic>=0.25",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
]

[project.scripts]
warepact = "warepact.cli.main:app"

[project.urls]
Homepage = "https://github.com/mfbatra/warepact"
Documentation = "https://docs.warepact.dev"
Repository = "https://github.com/mfbatra/warepact"
Issues = "https://github.com/mfbatra/warepact/issues"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.mypy]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = [
    "yaml",
    "snowflake.*",
    "redshift_connector",
    "psycopg2.*",
    "google.*",
    "boto3",
    "botocore.*",
    "croniter",
    "fastmcp",
    "mcp.*",
    "databricks.*",
    "anthropic",
    "duckdb",
]
ignore_missing_imports = true

# Implementation packages: relax annotation-completeness rules.
# Core and interfaces remain fully strict.
[[tool.mypy.overrides]]
module = [
    "warepact.validators.*",
    "warepact.adapters.*",
    "warepact.parsers.*",
    "warepact.cli.*",
    "warepact.mcp.*",
    "warepact.plugins.*",
]
disallow_untyped_defs = false
disallow_untyped_calls = false
disallow_any_generics = false
disallow_subclassing_any = false
warn_return_any = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
